feat: Add securityAndAnalysis to lateInitializer
Crossplane will fetch the state of the real world before updating the resource. If repository transistions
from internal to public it will read the state of securityAndAnalysis and merge that with what is set
from the managed resource. Github does not allow `.securityAndAnalysis.advancedSecurity` to be set when
public so this causes problems going form internal to public.

closes: https://github.com/coopnorge/provider-github/issues/34
ref: https://github.com/crossplane/upjet/blob/main/docs/configuring-a-resource.md#further-details-on-late-initialization